+++ packages/synchro-charts/src/components/charts/sc-bar-chart/sc-bar-chart.tsx
@@ -23,7 +23,7 @@ import { DEFAULT_CHART_CONFIG } from '../sc-webgl-base-chart/chartDefaults';
 import { RectScrollFixed } from '../../../utils/types';
 import { Trend } from '../common/trends/types';
 import { DATA_ALIGNMENT } from '../common/constants';
-import { validate } from '../../common/validator/validator';
+import { validate } from '../../common/validator/validate';
 
 // The initial size of buffers. The larger this is, the more memory allocated up front per chart.
 // The lower this number is, more likely that charts will have to re-initialize there buffers which is
