+++ packages/synchro-charts/src/components/sc-kpi/sc-kpi.tsx
@@ -3,7 +3,7 @@ import { Component, h, Prop } from '@stencil/core';
 import { DataStream, MessageOverrides, MinimalViewPortConfig } from '../../utils/dataTypes';
 import { RenderCell } from '../sc-widget-grid/types';
 import { Annotations, ChartConfig } from '../charts/common/types';
-import { validate } from '../common/validator/validator';
+import { validate } from '../common/validator/validate';
 
 const MSG =
   'This visualization displays only live data. Choose a live time frame to display data in this visualization.';
