+++ packages/synchro-charts/src/components/sc-widget-grid/sc-widget-grid.tsx
@@ -15,7 +15,7 @@ import { DATA_ALIGNMENT } from '../charts/common/constants';
 import { isMinimalStaticViewport } from '../../utils/predicates';
 import { parseDuration } from '../../utils/time';
 import { getDataStreamForEventing } from '../charts/common';
-import { validate } from '../common/validator/validator';
+import { validate } from '../common/validator/validate';
 
 const MSG =
   'This visualization displays only live data. Choose a live time frame to display data in this visualization.';
